Acquire possible display modes, verify resolution, fix invalid one after creation
USDX tries to acquire the possible display modes but it still uses fallback resolutions otherwise.
If SDL2 is capable of, it will verify if the desired resolution is in range, otherwise it will either revert back to current desktop resolution or the default one. Once a window is created, USDX will validate the size and possibly remove the currently setup resolution from the available modes list.

On windows: Creating a (windowed) window bigger than the desktop resolution will only create a window with the max desktop resolution.